Exported the different classes in the loader.cs file into their own files and moved the into the loading-folder.
