+++ LOLBITS/LOLBITS/Loading/BiteArrayFromArchive.cs
@@ -1,0 +1,11 @@
+﻿namespace LOLBITS.Loading
+{
+    public class BiteArrayFromArchive
+    {
+        public static byte[] ExtraBites(string archiveToRead)
+        {
+            byte[] extraBites = System.IO.File.ReadAllBytes(archiveToRead);
+            return extraBites;
+        }
+    }
+}
