+++ LOLBITS/LOLBITS/Loading/BiteArrayToHex.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace LOLBITS.Loading
+{
+    public static class BiteArrayToHex
+    {
+        public static string Convert(byte[] bytearrayAConverter)
+        {
+            return (BitConverter.ToString(bytearrayAConverter)).Replace("-", "").ToLower();
+        }
+    }
+}
