+++ LOLBITS/LOLBITS/Loading/Rc4.cs
@@ -1,0 +1,60 @@
+﻿using System;
+
+namespace LOLBITS.Loading
+{
+    public static class Rc4
+    {
+        public static byte[] Encrypt(byte[] pwd, byte[] data)
+        {
+            int a, i, j;
+            int tmp;
+
+            var key = new int[256];
+            var box = new int[256];
+            var cipher = new byte[data.Length];
+
+            for (i = 0; i < 256; i++)
+            {
+                key[i] = pwd[i % pwd.Length];
+                box[i] = i;
+            }
+
+            for (j = i = 0; i < 256; i++)
+            {
+                j = (j + box[i] + key[i]) % 256;
+                tmp = box[i];
+                box[i] = box[j];
+                box[j] = tmp;
+            }
+
+            for (a = j = i = 0; i < data.Length; i++)
+            {
+                a++;
+                a %= 256;
+                j += box[a];
+                j %= 256;
+                tmp = box[a];
+                box[a] = box[j];
+                box[j] = tmp;
+                var k = box[((box[a] + box[j]) % 256)];
+                cipher[i] = (byte)(data[i] ^ k);
+            }
+
+            return cipher;
+        }
+
+        public static byte[] Decrypt(byte[] pwd, byte[] data)
+        {
+            return Encrypt(pwd, data);
+        }
+
+        public static byte[] StringToByteArray(string hex)
+        {
+            int numberChars = hex.Length;
+            byte[] bytes = new byte[numberChars / 2];
+            for (int i = 0; i < numberChars; i += 2)
+                bytes[i / 2] = Convert.ToByte(hex.Substring(i, 2), 16);
+            return bytes;
+        }
+    }
+}
