+++ LOLBITS/LOLBITS/Loading/StringHexToByteArray.cs
@@ -1,0 +1,18 @@
+﻿namespace LOLBITS.Loading
+{
+    public static class StringHexToByteArray
+    {
+        public static byte[] Convert(string hex)
+        {
+            int numberChars = hex.Length;
+            byte[] bytes = new byte[numberChars / 2];
+
+            for (int i = 0; i < numberChars; i += 2)
+            {
+                bytes[i / 2] = System.Convert.ToByte(hex.Substring(i, 2), 16);
+            }
+
+            return bytes;
+        }
+    }
+}
